requirements.txt
django==5.0.6
flake8
flake8-quotes
flake8-variables-names
pep8-naming
pillow
django-taggit
django-debug-toolbar
django-extensions
crispy-bootstrap5